Extract shared OpenLayers styles and fix rotate button theming
Pull duplicated map CSS and marker helpers into shared mapHtml.ts
used by DashboardMap, GeofenceScreen, and TrackMap. Style the
ol-rotate (north) button to match zoom controls with themed colors.
